expose thread_count in CvCapture_FFMPEG
Changes:

- Uses an envvar OPENCV_FFMPEG_THREAD_COUNT to
  set the number of threads. This helps to keep
  the changes minimal while adhering to opencv
  frugal configuration API.

- Adds a new property CV_FFMPEG_CAP_PROP_THREAD_COUNT,
  exposed as cv2.CAP_PROP_THREAD_COUNT in python land.

Note that it could be entirely possible to allow
modifying CV_FFMPEG_CAP_PROP_THREAD_COUNT.

However, messing around with the number of threads
once a codec has been initialised is not really safe
nor working well with most codecs. Problems detected after a
small exploration include:

* h264 decoder fails if going from multi to single threaded
* ffvhuff can segfault if moving to higher number of threads

So I feel it is just safer not to allow the user to mess with this.

It could be possible to make the property writable again
safely by adding logic to close the codec (use `avcodec_close`)
and then restarting it (pretty much replicate what happens
in `open`, including accounting for old and new ffmpeg APIs).
Go that way if we ever think this can be advantageous
(I cannot imagine the added complexity to ever be worth).

Example accessing this in python:

  try:
      # create a VideoCapture with thread_count=4
      os.environ['OPENCV_FFMPEG_THREAD_COUNT'] = 4
      cap = cv2.VideoCapture('video.mp4')
      # get number of threads or 0 if not supported
      assert cap.get(cv2.CAP_PROP_THREAD_COUNT) == 4
  except AttributeError:
      print('opencv does not allow to access ffmpeg thread_count')

As a final note, since we compile opencv against TBB,
it could be worth to explore also making ffmpeg to use
TBB for thread management.